suppAccountInsert - decimal number representation
Changed the data type for variable safAmount fro float to double.
Articles state should not use either data type to represent
currency although for this project should not be an issue.
The float type would round up when large numbers
e.g. 9999999.99 to 10000000.00

+++ cur/supcur.c
@@ -1172,7 +1172,7 @@ int suppAccountInsert()
   int rows;
   int val, upID, *params[1], length[1],  formats[1];
   int safActiveID, safSupID, safPrtID, safSupTypeID=0, safStartDt, safEndDt, safPayID, safProAcctID;
-  float safAmount;
+  double safAmount;
   char safSupAcctRef[30], safComment[30], safAlias[30];
   int cf;
   int newRec = 'y';
